+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 **Version 2.2.1.3018** *TBD*
 - Made it so the haptics don't keep getting triggered, if the slider is not moving.
 - The long-press gesture recognizer is now detected, anywhere on the screen.
+- The long-press slider will now continue, if "start immediately" is on.
 - Improvements in accessibility (voiceover, mostly).
 - Code cleanup.